added exception handler and exceptions to present services; Close #2

+++ src/main/java/com/example/finalprojectspring/service/AdminPageService.java
@@ -3,12 +3,14 @@ package com.example.finalprojectspring.service;
 import com.example.finalprojectspring.dto.UserEntityDTO;
 import com.example.finalprojectspring.entities.Role_Of_Users;
 import com.example.finalprojectspring.entities.UserEntity;
+import com.example.finalprojectspring.exeption.ApiRequestExeption;
 import com.example.finalprojectspring.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class AdminPageService implements IAdminPageService {
@@ -31,6 +33,14 @@ public class AdminPageService implements IAdminPageService {
 
     @Transactional
     public void deleteUserByEmail(String email){
+
+        final Optional<UserEntity> existenceUserByEmail =
+                Optional.ofNullable(userRepository.findByEmail(email));
+
+        if (existenceUserByEmail.isPresent()){
+            throw new ApiRequestExeption("This user do not exist or already deleted");
+        }
+
         userRepository.deleteByEmail(email);
     }
 }
